buffers
git-svn-id: svn://elaine/murooma/trunk@182 d8a302eb-03bc-478d-80e4-98257eca68ef

+++ client/stream.h
@@ -29,7 +29,7 @@ private:
 	void resetBuffers();
 
 	long lastTick;
-	float sleep;
+	int sleep;
 	
 //	int correction;
 	Queue<std::shared_ptr<Chunk>> chunks;
